+++ src/Standards/SunAsterisk/Sniffs/Arrays/ArrayDeclarationSniff.php
@@ -1,4 +1,13 @@
-<?php
+<?php declare(strict_types=1);
+
+/*
+ * This file is part of the XuanQuynh package.
+ *
+ * (c) Quynh Xuan Nguyen <seriquynh@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 namespace XuanQuynh\CodeSniffer\Standards\SunAsterisk\Sniffs\Arrays;